backend/requirements.txt
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cohere==5.20.0
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
fastavro==1.12.1
filelock==3.20.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
groq==0.33.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.1.2
idna==3.11
jiter==0.11.1
joblib==1.5.2
numpy==2.3.4
openai==2.7.1
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
